test: add basic test for operator start
