+++ package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha 'src/**/*.test.js'",
+    "test:watch": "npm t -- --watch",
+    "posttest": "npm run lint",
     "lint": "eslint . --config .eslintrc",
     "format": "prettier --write .",
     "check-format": "prettier --check .",
@@ -18,6 +20,7 @@
     "lodash": "^4.17.15"
   },
   "devDependencies": {
+    "chai": "^4.2.0",
     "eslint": "^7.0.0",
     "eslint-config-emarsys": "^5.1.0",
     "eslint-config-prettier": "^6.11.0",
@@ -26,6 +29,7 @@
     "eslint-plugin-no-only-tests": "^2.4.0",
     "eslint-plugin-prettier": "^3.1.3",
     "eslint-plugin-security": "^1.4.0",
+    "mocha": "^7.1.2",
     "prettier": "^2.0.5"
   }
 }
