+++ src/index.js
@@ -1,19 +1,3 @@
-const Operator = require("./operator");
+const start = require("./start");
 
-const main = async () => {
-  const operator = new Operator();
-  await operator.start();
-
-  console.log("Secret operator running");
-
-  const exit = () => {
-    operator.stop();
-    process.exit(0);
-  };
-
-  process
-    .on("SIGTERM", () => exit("SIGTERM"))
-    .on("SIGINT", () => exit("SIGINT"));
-};
-
-main();
+start();
