+++ src/start.test.js
@@ -1,0 +1,9 @@
+const start = require("./start");
+
+describe("operator/start", () => {
+  it("starts without errors", async () => {
+    const operator = await start();
+
+    operator.stop();
+  });
+});
