+++ src/test-setup.js
@@ -1,0 +1,3 @@
+const chai = require("chai");
+
+module.exports = { expect: chai.expect };
